Fix case sensitivity: rename Components to components
- Resolved Git casing conflict between src/Components/ and src/components/
- All imports now consistently use lowercase components folder
- Fixed TypeScript errors related to case-sensitive file paths
