+++ src/components/activity-detail-components/Before.jsx
