+++ src/components/activity-detail-components/HeaderSection.jsx
