+++ src/components/activity-detail-components/LocationCard.jsx
