+++ src/components/activity-detail-components/ParticipantCard.jsx
