+++ src/components/profile_components/RecentActivities.jsx
