+++ src/components/styled/MiddleSection/Middle.styled.jsx
