+++ src/components/styled/act-detail-style-comp/Before.styled.jsx
