+++ src/components/styled/act-detail-style-comp/Common.jsx
