+++ src/components/styled/act-detail-style-comp/HeaderSection.styled.jsx
