+++ src/components/styled/act-detail-style-comp/HostCard.styled.jsx
