+++ src/components/styled/act-detail-style-comp/LocationCard.styled.jsx
