+++ src/components/styled/act-detail-style-comp/Participants.styled.jsx
