+++ src/components/styled/act-detail-style-comp/TitleCard.styled.jsx
