+++ src/components/styled/activity-detail-comp/ActivityDetail.styled.jsx
