+++ src/components/styled/activity-detail-comp/Middle.styled.jsx
