+++ src/components/styled/feed-style-comp/Card.styled.jsx
