+++ src/components/styled/feed-style-comp/Feed.styled.jsx
