+++ src/components/styled/feed-style-comp/Footer.styled.jsx
