+++ src/components/styled/feed-style-comp/Middle.styled.jsx
