+++ src/components/styled/feed-style-comp/NavBar.styled.jsx
