+++ src/components/styled/profile-style-comp/ActivityStatus.styled.jsx
