+++ src/components/styled/profile-style-comp/Cover.styled.jsx
