+++ src/components/styled/profile-style-comp/Interest.styled.jsx
