+++ src/components/styled/profile-style-comp/RecentActivities.styled.jsx
