+++ src/pages/Login.jsx
@@ -2,9 +2,9 @@ import React from "react";
 import {
   PageWrap,
   Split,
-} from "../Components/styled/login-style-comp/Login.styled.jsx";
-import LoginLeft from "../Components/login-components/LoginLeft.jsx";
-import LoginRight from "../Components/login-components/LoginRight.jsx";
+} from "../components/styled/login-style-comp/Login.styled.jsx";
+import LoginLeft from "../components/login-components/LoginLeft.jsx";
+import LoginRight from "../components/login-components/LoginRight.jsx";
 const Login = () => {
   return (
     <PageWrap>
